Rewrite undo and discard worker implementation
To improve the efficiency of the rollbacks, we create three queues and a
hash table for the rollback requests.  A Xid based priority queue which will
allow us to process the requests of older transactions and help us to move
oldesdXidHavingUndo forward.  A size-based queue which will help us to
perform the rollbacks of larger aborts in a timely fashion, so that we don't
get stuck while processing them during discard of the logs.  An error queue
to hold the requests for transactions that failed to apply its undo.  The
rollback hash table is used to avoid duplicate undo requests by backends and
discard worker.

To process the request, we get the request from one of the queues, search
it in hash table and mark it as in-progress and then remove from the
respective queue.  After that, we perform the request and remove it from
hash table.

Undo launcher is responsible for launching the workers iff there is some
work available in one of work queues and there are more workers available.
The worker is launched to handle requests for a particular database.  Each
undo worker then start reading from one of the queue the requests for that
particular database.

The discard worker is responsible for discarding the undo log of
transactions that are committed and all-visible or are rolledback.  It also
registers the request for aborted transactions in the work queues.  It
iterates through all the active logs one-by-one and try to discard the
transactions that are old enough to matter.

Designed by: Andres Freund, Amit Kapila, Robert Haas, Thomas Munro
Author: Amit Kapila, Kuntal Ghosh, and Dilip Kumar