Fix an issue in UndoDiscardOneLog
While calculating xmin in DiscardWorkerMain, we ignore the vacuum and
autovacuum xids. But, in zheap, a vacuum/autovacuum process can write
undo records that needs to be discarded.  So, we can't ignore the
autovacuum/vacuum process while calculating the xmin.  Else, we may
consider an in-progress vacuum/autovacuum xid as aborted which is wrong.

As of now, when a backend performs VACUUM, we forcefully clear
the vacuum flag from MyPgXact in lazy_vacuum_zheap_rel.  Hence, the
problem comes only with autovacuums.  We've to investigate the correct
way to handle this situation. For now, I've made the check for
registering a rollback request more strict.  It checks for the following
conditions to decide whether a xid is aborted:
1. xid didn't commit.
2. xid is not in progress. (added in this commit)
3. the corresponding undo record is not being rolled back.
4. the corresponding database is still exists.

Reported by Andres Freund. Investigated and Patch by me. Reviewed by Amit
Kapila and Dilip Kumar.