Fixed toast rollback related bug in zheap_undo_actions
In zheap_undo_actions, we were locking the relation based on rellock
flag.  We assume that if our current backend aborts and then in applying
undo actions, we have all locks of this transaction but for toast relation
we don't have lock because we acquires and release the lock on relation in
ztoast_save_datum.  So we were getting assert failure of relation lock.

As suggested by Robert and Andres, I removed rellock flag and now we
are locking the relation unconditionally.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila