Fix a bug in undoworker
When we remove an element from the binary heap, we basically swap that
element with the last one.  So, when we're removing multiple elements
from binary heap using a loop, once we remove an element from a
location, we need to examine the same location again.

Reported by Mahendra Singh Thalor