Move more logic into ZGetMultiLockMembersForCurrentXact.
It is renamed to ZCurrentXactHasTupleLockMode.  The loop performed
by every caller to check whether a high-enough lock mode is present
is now moved into the function itself.  This not only takes less
code but also permits us to stop early if we find that the lock
mode is present.  It also saves allocating and deallocating memory
unnecessarily.