Use common routine to lock the tuple.
During update we lock the tuple when we momentarily release the page lock.
However, to lock tuple we were using a code very similar to what we have in
zheap_lock_tuple_guts.  This is probably done this way by seeing heap, but
here we have somewhat more code, so it is better to use a common routine.

While doing this refactoring, I have noticed that zheap_lock_tuple_guts
accepts too many parameters which we should try to reduce, but that is a
matter of separate commit.

Suggested by Andres Freund, Patch by me, reviewed by Kuntal Ghosh