Fixed a bug in rollback handling.
In zheap_undo_actions, when we try to open relation, it is possible
that relation already dropped by that time.  In such case, we were not
able to open relation and we were throwing an error.   Now, the
request is again registered in the error queue for processing which is
not correct.  We must skip processing the undo actions in this case.
Also, we need to abort the transaction in case of error and clean the
error state.

Patch by Kuntal Ghosh and me, reviewed by Amit Kapila