Fixed bug for multi-locker clearing bit
In ZIsAnyMultiLockMemberRunning, we were checking that if there
is any running transaction or not, if not, then we were clearing
multi-locker bit.  But it is possible that transaction is aborted and
undo actions still not applied, then we can't clear bit.
So as suggested by Amit and Dilip, now we are applying actions for
aborted transaction in ZIsAnyMultiLockMemberRunning.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila