Fixed a bug in UndoSetPrepareSize
In UndoSetPrepareSize, if nrecords were less than or equal to
MAX_PREPARED_UNDO, then we were not calling UndoRecordAllocate,
so we were not setting meta.  So we were logging invalid meta.
Due to this, in redo, we were getting segment fault in
undolog_xlog_meta.  Fixed this.

Patch by Mahendra Singh Thalor, reviewed by Dilip Kumar