Added isolation test case for commit 3bacab225c6c88b78e4
When we were trying to update a tuple and concurrently tuple is
deleted and item id is marked as deleted, then we were getting
segment fault in ZHeapDetermineModifiedColumns because tuple data
is setted to null in ZHeapTupleSatisfiesUpdate for deleted item id.
So added test case for this failure.

Patch by Mahendar Singh Thalor, reviewed by Amit Kapila