Defect and enhancement in multi-log support
Patch fixes different problems for identifying the undo log switch during
a transaction.
a) Earlier if a log switch is detected but if there is an error and that undo is
never inserted then if there is next undo record from the same transaction then
the log switch is never detected.  Now, we have kept in the TopTransactionState
so the first successful undo record after the log switch will keep the log switch
header.
b) During recovery we are writing extra log switch WAL to detect the log switch.
In this patch instead of writing extra WAL we are identifying the buffers registed
by the main WAL.