Fixed bugs in zheap_lock_updated_tuple to free tuple and to lock next tuple
In zheap_lock_updated_tuple, after locking buffer, we are fetching
tuple again and before that we are freeing tuple.  So if itemId is
deleted then we will not fetch tuple but we were trying to free it.
So fixed this.
Without setting zhtup info, we were using it, so fixed this.

Patch by Mahendra Singh Thalor, reviewed by Kuntal Ghosh and Amit
Kapila