Replaced TransactionIdDidAbort check by not-commit and not-in-progress in 5 places to fix some bugs.
We were using TransactionIdDidAbort to identify aborted transaction,
but if there was a transaction is running before crash, then after
restart, this check will not consider that transaction as aborted.
So changed this check to support such cases by replacing with
not-commit and not-in-progress check.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila, Kuntal
Ghosh and Dilip Kumar