Fixed bug in zheap_lock_tuple for SKIP lock case
In zheap_lock_tuple, if result is not TM_Ok, we are assuming that
this itemId is either deleted or tuple is modified but in case of
LockWaitSkip, it is possible that itemId is normal and tuple is not
modified.
For example, if there was one running transaction, then for that xid,
we try conditional wait(LockWaitSkip).  If we couldn't get success,
then we return TM_WouldBlock.  But it is possible that after our
conditional wait, transaction can abort and rollback actions can
applied.  So itemId can be normal for this case.  So changed assert
to handle this case.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila