Fixed a bug in process_and_execute_undo_actions_page to skip already applied undo actions
In process_and_execute_undo_actions_page, we are calling UndoRecordBulkFetch
to collect records from urec ptr but it is possible that actions are
already applied but fxid is not cleared from slot.  In that case, we
might fetching records of other transactions and applying actions that
was wrong.  So fixed this by comparing fxid of slot and first uur_fxid.
Instead of this check, we can pass fxid to UndoRecordBulkFetch to collect
records but for code simplicity, we are not changing UndoRecordBulkFetch
api.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila