Fixed bug to traverse non in-place updated tuple chain to lock latest version
In ZHeapTupleFetch, if we find that tuple is non in-place updated
and that transaction is committed and tuple is not visible, then we
were not fetching tuple and new ctid.  So on, we were not able to
traverse chain as we found itemId as deleted and we stop there by
assuming end of chain.  So fixed this by passing flag keep_tup.  If
keep_tup flag is set, means caller need invisible tuple and ctid to
traverse chain.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila and Kuntal
Ghosh