Fixed bug in GetTupleForTrigger to fetch tuple
In GetTupleForTrigger, if we couldn't fetch tuple with passed snapshot,
then we are trying to fetch tuple with latest snapshot.  But in zheap,
if we couldn't fetch tuple in first attempt, then tid might get overwritten
with the latest tid in non-inplace update case.  So make the copy of tid
before fetching the tuple.  So that in the unsuccessful case if we need
to re-fetch the tuple then we have the right tid.

Patch by Mahendra Singh Thalor, reviewed by Dilip Kumar