Fixed bug in RefetchAndCheckTupleStatus for single_locker_xid
In RefetchAndCheckTupleStatus, if tuple is marked as xid lock only
and even single_locker_xid was as invalid, then we were trying to
get current single_locker_xid by calling GetLockerTransInfo.  It is
quite possible that single locker is aborted but actions are not
applied so we are always getting same locker xid till actions are
not applied.  So fixed this.
Note: This is a temporary fix.

Patch by Mahendra Singh Thalor, reviewed by Kuntal Ghosh