Fixed a bug in zheap_lock_updated_tuple to reverify xwait
To lock tuple, we wait for in-progress transaction to commit/abort
but it is possible that after our wait, some other backend can lock
tuple and perform non in-place update before we acquire lock on tuple
so we should reverify infomask of tuple and we should calculate xwait
again.
NOTE: This is a temporary fix.  Ideally, we should verify xwait in
ValidateTuplesXact.

Patch by Mahendra Singh Thalor, reviewed by Amit Kapila