Automatically close UndoRecordSets on xact or subxact abort.
Along the way, I made a few other, mostly-related changes.
UndoPrepareToMarkClosed now returns a boolean; if the return value is
false, you have the option to skip marking the UndoRecordSet closed,
saving WAL in cases where we didn't actually write anything.

Also, there are now separate functions UndoRelease, which only
releases buffer locks and pins, and UndoDestroy, which also destroys
the UndoRecordSet. I think that's safer than relying on the internal
state of the UndoRecordSet to know which operation the caller
intended, and it allows for better debugging cross-checks.

Currently, xact commit, subxact commit, and prepare don't use
this new infrastructure; that should be fixed.