Close UndoRecordSets on xact or subxact commit or preapre.
Also, fix some bugs in the previous commit. Turns out <= and >=
are not interchangeable.

Also, make the UndoRecordSet objects created by foo_create()
transactional, so that they are automatically closed at commit
or abort time. Perhaps it's useful to have the nontransactional
version available also, but for now I'll bet this is more
useful.