Clean up a few details around persistence handling.
Rename NPersistenceLevels to NUndoPersistenceLevels since, at least
for now, this is defined in an undo-specific header.

Add named constants for the undo persistence levels, because some
code I want to merge needs them. Put the definitions in undodefs.h
so that xactundo.h can avoid including the unwieldly undolog.h
and all of its dependencies.

Replace the macro implementing a trivial perfect hash with an
function that does the same thing in a stupider but perhaps more
robust manner. It can survive having a RELPERSISTENCE_* constant
renamed and will fail an assertion if a new one is added and the
new value shows up here.

Move that function out of line, too, so that we don't have to include
even more headers in undolog.h.  I'm not sure that undolog.c is the
right place for this either, but there's nothing better right now.