Refactor undo insertion code to use per-buffer flags.
The previous code could call XLogRegisterBuffer more than once for
the same buffer if undo was inserted and then recordset was closed
in the same WAL-logged operation. That has the effect of clobbering
data previously registered for the buffer, so instead, use a flag
to say whether or not that's still needed.

(This kind of approach might also be better for making this code
worked with unlogged tables, which will need to skip all of the
XLogRegister stuff, but it would need more hacking.)

The previous code also called PageInit at the wrong time -- when
preparing the operation, rather than when performing it. Fix that
using another flag, indicating whether or not PageInit is needed.

Also use a third flag to track whether MarkBufferDirty is needed.
Unlike the previous two changes, this doesn't fix anything directly,
but it seems like the right thing to do for symmetry; e.g. it lets
us get rid of UndoInsertState's last_buffer_index member.

Along the way, put back UndoRecordSet's need_type_header member,
which I removed yesterday. The replacement coding broke WAL record
assembly by changing the value of bytes included in the buffer data
before XLogInsert() was performed.