Fix incorrect offset/size calculations in UndoMarkClosed.
When the length word of an undo record set chunk spans a page
boundary, the number of bytes that need to be updated on the
second page is equal to the number that didn't fit on the first
page, not the number that did fit.

Moreover, those bytes are located at the beginning of the second
page, not the end.