Refactor UndoMarkClosed.
Move the per-page work to a separate function to avoid having two
copies of the code, one for the main page and one for the
continuation page.

Along the way, create a named constant for the size of an individual
chunk_header_ops item, and use it in a few existing places and one
new one instead of writing "4".

Only call XLogRegisterBufData in this code path if the persistence
level is "permanent". This still isn't enough to make unlogged and
temporary relations work, but it's another step in that direction.