New undo xlog buffer data format.
Modifications to undo pages are now recorded in a slightly more
scrutable format, with tools to encode and decode them, and are
displayable with pg_waldump.

As part of that, add the insertion pointer to every registered
buffer, rather than relying on FPIs in strange new ways.

Skip updating pages when they are not found (presumed discarded
by later WAL) or when they were restored from FPI, but still do
the maths required to finish up with the right shm insert
pointer.  For now that's only needed once we reach consistency,
due to having insert pointers store in bufdata for every undo
insertion, but the goal is to get back to dropping them sometimes,
so we need to be able to maintain them even when skipping the
updates.

Multi-chunk URSs will now be closed incrementally, so the number
of buffers that must be registered with any given WAL record is
anchored to the ground (work in progress).

Make a single function UndoReplay() that handles both insertion
and closing (if there was any) during recovery; client code needs
to call it with NULL record data if something might have been
closed at DO time, and with non-NULL record data if there was
a call to UndoInsert() at DO time, plus any amount of closing.

Use the new undopage.c code that maintains the new undo page
header format.

Introduce UndoBuffer struct, to keep flags together with buffers
in an array.

WORK IN PROGRESS -- MULTIPLE THINGS NOT WORKING YET