Add UndoXLogRegisterBuffers().
This replaces some ad-hoc code that tried to register buffers and
bufdata the right number of times over several operations that will
be redone by a single WAL record.  This must be called exactly once
for every dirty URS that before the WAL record is inserted.

Not yet done: adding the extra calls to xactundo.c.