hackaround: Kinda fix transaction state handling in PerformBackgroundUndo().
Before we'd fail the second time PerformBackgroundUndo() is run inside
the same worker, because of unexpected transaction states. Two
problems:
1) CommitTransaction() sets the state to TRANS_ABORT, instead of the
   "normal" TRANS_DEFAULT when the signalling undo processing. Which
   is not an accepted state to start a transaction. Leading to
   failures in the next transaction.
2) CommitTransaction() does not actually clean up the transaction, so
   we fail trying to start the next transaction, even if fixing up 1).

Try to mimick normal transaction state transition. But it's probably
not the right way to do things - for one, several of these operations
can fail, and the transaction state machinery shouldn't get confused
by that.