undoxact: Add undoxacttest test module.
This basically just provides a way to simulate interactions with the
xactundo layer, without needing a full blown table AM.

The idea is to just modify a counter - a single one for now, but quite
possibly more later - stored inside a heap table. That way
comparatively little scaffolding is needed to allow for testing, and
introspection.

A following commit will add a test file using this.