Fix some problems in undopage.c.
UndoPageInsertHeader and UndoPageInsertRecord should not Assert that
the data doesn't overrun the page, because they're smart enough to
handle the case where it does by writing only the part that fits.

Adjust the comment for UndoPageInsertHeader to clarify that it is
not necessary for type_header to be NULL when type_header_size == 0.

Fix file name in file header comment.