undorequest: Install an explicit state representation.
This makes it possible to loop over every UndoRequest in the system
and understand the state of each one easily. Before, there was no
real easy way to understand this. Objects had states, but you
couldn't necessarily figure out what state a given object was in
just by examining it.