undorequest: Maintain a list of used UndoRequest objects.
This makes it more efficient to iterate over the used objects,
without having to traverse all of the free ones. Or at least,
it should be more efficient if, as is expected, the number of
used objects is normally small compared to the total number of
objects.

Along the way, use dlist instead of an open-coded singly linked
list.