undorequest: Move MarkBufferDirty() calls.
Previously, they were in UndoXLogRegisterBufferData(), which had
the advantage of doing it only once for each page touched.  It's not
clear from transam/README that that was wong, since it was still
before XLogInsert(), but it's certainly the usual convention to
mark the buffer dirty close to the places that actually write to
the page, so let's just do that.