undorecordset: Incrementally mark chunks closed.
In earlier code, we closed all the chunks with one WAL
record.  For example, the COMMIT would update all the
chunk headers with lengths.  That required practically
unbounded buffer registrations.  Now, a single WAL record
that inserts a new undo record in a new chunk will carry
bufdata instructions to mark the previous chunk closed
(that is, update its header with a length).