Fix up dangling undo record sets at startup.
If we crashed, then we might have left some undo record sets that have
a length of zero in their final chunk.  We need to find those, for two
reasons:

* we need to set their length, so that we'll know how do discard them
* we need to be able to tell URS type owners such as xact.c that a
  URS they created is being automatically closed, so that they can
  act on that (for example, queue an undo request to triggerarollback)

Also fix a pre-existing problem that bufdata was being dropped when
FPWs were logged (because I didn't use REGBUF_KEEP_DATA).

There are several TODO notes and questions remaining in this code,
and I'm not at all sure that this is the right place in StartupXLOG();
more testing and study required.