Properly reassociate GlobalTransactions with UndoRequests after restart.
This allows COMMIT PREPARED and ROLLBACK PREPARED to work properly
with undo even after a restart.

There are still going to be problems if a crash occurs, but that's
also true, at present anyway, even when no prepared transactions
are involved. So, AFAIK, this makes prepared transactions work just
as well as anything else does currently.