Tell URS owners when their URSs are closed (take II).
Call XactUndoHandleCloseURS() whenever a URS is closed by
CloseDanglingUndoRecordSets() or in recovery.  Give it
the full URP range and type-specific header (which holds
the fxid).

In order to be able to do this in recovery without consulting
data pages, attach the above to the URS_XLOG_CLOSE instruction
in the bufdata for the undo page when the final chunk is closed.

Other small improvements:
* remove a fixed-sized buffer that arbitrarily limited header size
* improve some variable naming in passing
* fix uninitialized chunk header
* fix recovery problems found by binary comparison of pages