Draft some code for now-renamed XactUndoCloseRecordSet.
This turns out to be a bit problematic, becaue we need some
information the caller doesn't currently provide. I rejiggered
things so that we can find out whether we've got a COMMIT,
an ABORT, or a PREPARE, which helps.

But I still don't have the database OID or the total amount
of permanent undo for the record set, so this isn't really
right yet.