undolog: Merge UndoLogGetSlot with find_undo_log_slot.
These functions are basically identical, except that
UndoLogGetSlot is non-static and takes a missing_ok flag.

Since some of the previous callers of find_undo_log_slot want
the missing_ok = false behavior, this actually saves a few lines
of code.

I mostly kept the comment from UndoLogGetSlot rather than the
one from find_undo_log_slot because it seemed a bit shorter
and clearer, but that might not be optimal.