xactundo: Add dboid to record set header; adjust XactUndoCloseRecordSet.
This probably doesn't totally work yet, but it's probably an
improvement.

At the moment, recovery fails when it sees a close-URS attached
to an XLOG_NOOP record, because the code added by my commit
64481ab2c8c3fb853d08bf1dc587642d0c831b6d expects to see that only
with COMMIT, ABORT, or PREPARE. That needs to be fixed before I
can really debug this.