Make unlogged and temporary undo logs work as expected.
Reset temporary undo logs at startup.  Do the same for unlogged undo logs if
there is a crash restart.  That matches the conditions under which temporary
and unlogged relations are reset.

For now, resetting means blowing away all segment files and fastforwarding
the discart and insert pointer to the next segment boundary.  We should
probably learn to reuse the whole undo log's address space, but that needs
more work.