Paritally restore undo log tablespace support.
In the "UndoRecordSet" redesign, the tablespace support was left in a broken
state.  This commit brings it back, though more work is needed.  The user
interface is the same as before:

* you can set undo_tablespaces to a list of tablespace names
* undiscarded data prevents a tablespace from being dropped
* tablespace names aren't resolved immediately

The new coding is a little more explicit about what's going on, adding a new
"state" field to UndoLogSlot so that we know whether it's currently in use.
An undo log (and thus an UndoLogSlot that manages it) can be in one of three
locations:

* on a shared free list
* on a private free list (previously known as "sticky" or "attached")
* actively being used in an UndoRecordSet

Only the last category prevents the containing tablespace from being dropped
(though it also needs to be empty).

Remaining problems:

XXX This could use some isolation tests or similar to test various cases in the
locking protocol vs concurrent DDL.  Probably need to make it easier to deal
with tablespaces in regression tests first (ie relative paths).

XXX Currently, the error unwinding code blows up if you set undo_tablespaces to
an unknown tablespace name, after an ERROR is raised; needs to be investigated.

XXX The private freelists are actually not lists, but just a single item at the
moment, but we know we want to be able to have fast access to at least two in
some cases, so we can make multixacts fast.