xactundo: Account for failures while inserting undo.
If the transaction fails after getting an UndoRequest but before
actually inserting any undo, the old code would fail an assertion,
because it didn't want to perform an amount of undo equal to zero.

Fix that by deallocating the UndoRequest if we have one but there's
nothing in it.

Per report from Thomas Munro.